fix(nftnl): handle duplicate ACK from kernel during module autoload
A kernel bug was introduced in commit bf2ac490d28c which makes it so
that a duplicate ACK is returned when a module is auto-loaded.
Additionally, it seems that the maintainers of nfnetlink do not
recommend tracking the number of ACKs which were sent and received.

This change makes it so that no assumptions are being made on the number
of messages being returned from the kernel.